requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
httpx>=0.24.0
python-dateutil>=2.8.2
